refactor: move files from src to src/web
