--- src/src/app/api/layout.tsx
@@ -1,7 +1,0 @@
-export default function DocsLayout({ children }: any) {
-  return (
-    <section>
-      {children}
-    </section>
-  )
-}
