--- src/src/app/page.tsx
@@ -1,5 +1,0 @@
-export default function Home() {
-  return (
-    <h1>Hello World</h1>
-  )
-}
