+++ src/web/package.json
@@ -34,8 +34,7 @@
     "@types/react": "18.2.22",
     "@types/react-dom": "18.2.7",
     "@types/swagger-ui-dist": "^3.30.1",
-    "prisma": "^5.3.1",
-    "vercel": "^32.3.0"
+    "prisma": "^5.3.1"
   },
   "prisma": {
     "schema": "src/shared/database/schema.prisma"
