+++ src/web/src/app/api/layout.tsx
@@ -1,0 +1,7 @@
+export default function DocsLayout({ children }: any) {
+  return (
+    <section>
+      {children}
+    </section>
+  )
+}
