+++ src/web/src/app/page.tsx
@@ -1,0 +1,5 @@
+export default function Home() {
+  return (
+    <h1>Hello World</h1>
+  )
+}
